requirements.txt
transformers
datasets
pandas
evaluate
seqeval
torch
gliner
torchvision
torchaudio
accelerate
huggingface_hub
scikit-learn
numpy
pytorch-crf
python-dotenv
gliner2